qemu: Bump qemu mainline version to 2.6
Bump the mainline version of QEMU to 2.6 as that is required to compile with
Xen 4.7.

This is an exact copy of the files from current mainline meta, except for
changing the qemu-targets path.
Taken from commit: 087c580b286816265f487e02746bfa6e26081554

This should not make any changes to qemu-native.

Signed-off-by: Alistair Francis <alistair.francis@xilinx.com>

Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>