smartpm: Fix for unknown error when installing packagegroup-petalinux-x11
Installing x11 packagegroup through smart results in "error:Unknown
error" message. This patch is a fix for the error message.

SmartPM misinterprets empty problems list as a valid problem list.  This
is because the problems list is checked against None where as in some
cases the problems list is empty instead of having a value of None.

This patch fixes this issue by checking for empty problems list too.

Signed-off-by: Bhargava Sreekantappa Gayathri<bhargava.sreekantappa-gayathri@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>