libvcu-xlnx_git.bb: Compile from custom vcu repo
Compile VCU libs and bins from custom repo and install them in rootfs.

Install vcu control SW headers needed for -dev packages and this are
also needed for compilation by libomxil-xlnx recipe.

Set compatible machine as zynqmp as this is the only architecture
supporting VCU with which the VCU stack provided by this recipes are
validated.

Signed-off-by: Devarsh Thakkar <devarsht@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>