petalinux.conf: Update libmetal and OpenAMP preferred version
As upstream 2017.04 is already released, and Xilinx 2017.1 is still
using 2017.04 version. For any changes since 2017.04 should be put in a
new version. meta-openamp already has 2017.10 bb to use the tip of the
upstream. We should use 2017.10 version instead of just update the
SRCREV of 2017.04 version as this will confuse user.

Signed-off-by: Wendy Liang <jliang@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>