packagegroup-petalinux-matchbox: Adds icon theme dependency
When running a petalinux image that uses matchbox, the terminal
icon isnt shown, and some warnings are shown at boot log.

The reason for this is that the matchbox packagegroup
does not depend on the adwaita-icon-theme package,
resulting in the package manager deciding to install
a weak dependency to adwaita-icon-theme-symbolic, and while
that package does contain some small icons, it doesnt contain
an index.theme file, which is why matchbox complains.

The only reason this only happens on the terminal
application is because we dont install any other
applications by default, but if we did, the same thing
would happen for them.

This patch adds an RDEPENDS for the matchbox packagegroup
to the icon theme, causing the rpm file to require it and
making the package manager install it correclty when
executing do_rootfs

Signed-off-by: Alejandro Enedino Hernandez Samaniego <alejandr@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>
Signed-off-by: Bhargava Sreekantappa Gayathri <bhargava.sreekantappa-gayathri@xilinx.com>