packagegroup-petalinux-matchbox.bb: Update Package list
Similar to the packagegroup-petalinux-audio patch, if pulseaudio is in
Distro features, add packages pulseaudio-server,
pulseaudio-client-conf-sato, and pulseaudio-misc. This is a workaround
to allow gconf to save desktop configuration settings. Adding this line
to matchbox packagegroup too so that if audio packagegroup is not
enabled but pulse audio is in Distro features, gconf configurations will
still work.

Signed-off-by: Jaewon Lee <jaewon.lee@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>
Signed-off-by: Bhargava Sreekantappa Gayathri <bhargava.sreekantappa-gayathri@xilinx.com>