ultra96-startup-page.sh: kill ultra96 webapp flask server daemon
The ultra96 statup page script runs a dameon to start and stop the flask
server which launches the ultra96 webapp. When the flask server is
started the pid is stored in a FLASK_PID_NAME.pid file.  To kill the
server, the current script issues a kill command to FLASK_CMD.pid file
instead of FLASK_PID_NAME.pid file. As a result the init script does not
kill the webapp.

The current patch fixes the script by replacing the correct variable in
the kill command.

Signed-off-by: Sai Hari Chandana Kalluri <chandana.kalluri@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>