local.conf.sample: Add Vivado toolchain path
Vivado Design Suite is a software suite for synthesis and analysis of
HDL designs. Since, meta-xilinx-tools layer has introduced
vivado.bbclass, we need to specify XILINX_VIVADO_DESIGN_SUIT variable in
local.conf which will point to the Vivado installation directory.

This patch adds this variable to the local.conf.sample file which needs
to be modified to point to the user's Vivado installation direcctory

Signed-off-by: Bhargava Sreekantappa Gayathri <bhargava.sreekantappa-gayathri@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>