arduino-toolchain_1.0.bb: Drop custom arduino.mk file
For reset_stty, the arudino make file parses through different stty
stdin/out redirections such as -F, --file, -f etc. stty from core utils
prefers -F and --file redirection. The init daemon is unable to write 1
to the stty after it obtains a suitable redirection argument and break
the for loop in reset_stty. This is because daemon processes do not have
a controlling tty unlike a regular process. As a result we need to
explicitly set stty argument in the makefile so that it can be run by a
daemon process.

Signed-off-by: Sai Hari Chandana Kalluri <chandana.kalluri@xilinx.com>
Signed-off-by: Alejandro Enedino Hernandez Samaniego <alejandr@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>