gst-shark_0.2.1.bbappend: Add pkgconfig-native to dependency
Adding pkgconfig-native to dependency as this recipe
requires pkg-config. Without this patch, there is a configure error
because the script can't find pkg-config in recipe-sysroot-native. This
patch will require that as dependency so it will be there when
necessary. Change is being held in a bbappend now as we are using
upstream gst-shark recipe.

Signed-off-by: Jaewon Lee <jaewon.lee@xilinx.com>
Signed-off-by: Alejandro Enedino Hernandez Samaniego <alejandr@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>