sensor-mezzanine-examples:Modify touch sensor example:
In the current example, the touch sensor is connected on G3 GPIO pin. In
the current Ultra96 BSP GPIO pins E-L are being used by the hardware
desing to support MIPI Camera. The following pins are used by the
design:
    1. GPIO-G: Strobe        2. GPIO-H: ULPM
    3. GPIO-I: CAM1_RESET    4. GPIO-J: CAM1_PWDN
    5. GPIO-K: CAM2_RESET    6. GPIO-L: CAM2_PWDN

The above pins cannot be used by the sensor mezzanine adapter to drive
input, only output is possible. Hence the relay is connected to G2 while
the touch sensor is moved to GPIO pin AB

Update commitid for Ultra96-startup-pages; the touch and relay example
on the webapp is modified to reflect the above changes.

Signed-off-by: Sai Hari Chandana Kalluri <chandana.kalluri@xilinx.com>
Signed-off-by: Jaewon Lee <jaewon.lee@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>