run-postinsts: merge stderr and stdout when executing scripts
run-postinsts runs a given set of scripts during the first boot of the
device, when one of these scripts prints something to stdout, since
stdout is not available at that time, the script execution immediately
returns with an error, this error causes the script to terminate all
threads within the process, causing undesired behavior since the script
might still had to execute some other code.

For example, when running an update-alternatives script, which could
have several alternatives, when one of the update-alternatives
executions echoes something the script returns, and the rest of the
update-alternatives commands fail to execute.

We were only able to reproduce this in Petalinux, because it is missing
a ROOTFS_POSTPROCESS_COMMAND which creates logging variables for
run-postinsts scripts, when that is the case, it does not need stdout
hence the error is avoided.

Modify the run-postinsts script from oe-core, merging stderr to stdout
which should be harmless and avoids this issue.

Signed-off-by: Alejandro Enedino Hernandez Samaniego <alejandr@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>