rfdc: Cleanup of rfdc recipes to allow relative path for license file
Rfdc recipes were changing the S directory from ${WORKDIR}/git to a
deeper directory within ${WORKDIR}/git. Hence, LIC_FILES_CHCKSUM was set
using an absolute path (${WORKDIR}/git/license.txt). This was breaking
devtool and external source flows as the license file is not in that
absolute path. Changing LIC_FILES_CHCKSUM to use a relative path and
modifying rfdc recipes to be able to build and install without changing
the ${S} directory. This will make sure that LIC_FILES_CHECKSUM is found
for all workflows as long as the license file is in the top directory of
${S} (which is usually the case).

Signed-off-by: Jaewon Lee <jaewon.lee@xilinx.com>