recipes-apps: Remove sdfec example from meta-petalinux
- remove ./recipe-apps/sdfec-demo
- remove ./recipe-apps/sdfec-multi-ldpc-code
- remove ./recipe-libs/libsdfecuserintf
- add ./recipe-apps/sdfec/sdfec.bb

The patch removes the SDFEC example source code from meta-petalinux, due
to request to limit the size of included code in meta-petalinux.  The
sdfec.bb recipe file has been added which instructs meta-petalinux where
to find and how to build the SDFEC examples.  the SDFEC example storage
place is the new repository which is alocated for Linux examples

Signed-off-by: Dragan Cvetic <dragan.cvetic@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>