gstreamer1.0-plugins-*: rewire two patches modifying 'common' dir
gstreamer1.0* are packages that use a git repo inside another git repo (the
'common' directory in the upstream gstreamer1.0 package is included in the only
git source, but for our source, we are pulling 'common' separately from another
url.) For these cases of a git in another git devtool modify was failing
because the patches for the inner git are being applied but not committed
correctly and so we are getting the following error:

    On branch devtool
    Changes not staged for commit:
        modified:   common (modified content)

    no changes added to commit

Modifying the paths the two upstream patches changing 'common' are targeting
(i.e. from common/m4/introspection.m4 to m4/introspection.m4) and adding the
correct patchdir for each of the patches (common) also removing the upstream
patch gtk-doc-tweaks.patch that gstreamer1.0-plugins.inc is appending.

Signed-off-by: Jaewon Lee <jaewon.lee@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>