recipes-examples: Add recipe for 2019.2 AIE OOB
Added aie-matrix-multiplication Jupyter notebook example.This application
demonstrates square matrix multiplication that is accelerated on AIE using
standard matrix multiplication algorithm. The user can change the matrix size
and the number of the core utilized at compile-time. The speed-up for a matrix
size of 500x500 is 42x when compared with a naive implementation of matrix
multiplication on ARM Cortex-A72.

You can refer to the README for more information

Signed-off-by: Dishita Vashi <dishita.vashi@xilinx.com>
Signed-off-by: Manjukumar Harthikote Matha <manjukumar.harthikote-matha@xilinx.com>