xinput-calibrator: remove boot time auto launcher
This tends to mess up the X Server start up in some cases and ends up
generating wrong calibration data as well randomly.  We now remove it from the
default set up as the user already has an option to run this from the cmdline
or the SATO GUI once the system is up.

Xinput-calibrator need not be part of autolaunch as it might print Error logs
in scenarios where the display does not have touchscreen support. We can always
enable the support by modifing the config file to support calibration by
tweaking variable "HAVE_TOUCHSCREEN".

Signed-off-by: Awais Belal <awais_belal@mentor.com>
Signed-off-by: Alejandro Enedino Hernandez Samaniego <alejandro.enedino.hernandez-samaniego@xilinx.com>
Signed-off-by: Madhurkiran Harikrishnan <madhurkiran.harikrishnan@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>