jupyter-setup.sh: Print jupyter notebook URL to console on boot
During boot ensure the initscript prints the jupyter notebook URL to
console. Currently on boot, jupyter notebook URL is copied and stored in
/var/log/jupyter.log file instead of being displayed on the console. The
user is now forced to open the log file and copy paste the URL to the
browser. Update the initscript to print the URL to console so that user
doesnt have to open the log file.

Signed-off-by: Sai Hari Chandana Kalluri <chandana.kalluri@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>