tcf-agent: Upstream no longer appears to permit https downloads
Revert to the original tcf-agent_git.bb SRC_URI list.

Upstream git URL hasn't changed, but the http/https URL requires a new
unique URL.  There does not appear to be any reason we need the http
version, so use the git one that is stable instead.

Also use the upstream patches as well.  Only keep the changes that are
unique to this file.

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>