packagegroup-petalinux-gstreamer.bb: Include gstreamer plugins for zynqmp
Only gst-omx was etting selected for zynqmp due to gstreamer package group
variable was getting override for zynqmp to gst-omx as per commit
63e8610b87a0f848cd887352209f12ab481f1d02.

As the intention is to enable gst-omx only for zynqmp, we use append suffix
i.e. GSTREAMER_PACKAGES_append_zynqmp so that gst-omx will only get selected
for zynqmp (and not other platforms like everest).

Signed-off-by: Devarsh Thakkar <devarsh.thakkar@xilinx.com>
Signed-off-by: Sai Hari Chandana Kalluri <chandana.kalluri@xilinx.com>