image-xlnx-xen.bbclass:add class to generate xen boot files.
This patch adds support to generate the xen boot files using xen image-builder.
here image-builder is used to calculates the binary sizes and adds offsets and
generates xen_boot.scr and xen_boot.source files which is required for booting xen.

since it require generated images in output directory so this class is dependent with plnx-deploy.bbclass.
As xen is supported with only initrd images.so this class also based on initrd images.

Signed-off-by: Swagath Gadde <swagathg@xilinx.com>
Tested-by: Raju Kumar Pothuraju <raju.kumar-pothuraju@xilinx.com>
Signed-off-by: Jaewon Lee <jaewon.lee@xilinx.com>