packagegroup-petalinux-xen.bb: include qemu-system-i386 binary if vmsep distro feature is enabled
When vmsep distro feature is enabled, qemu packaging is modified and split into
qemu-<arch>, qemu-support, qemu-keymaps. This packaging is to ensure all of
the built qemu targets are not installed on target. Only the required qemu
binary can be installed.

For petalinux distro, only qemu-system-i386 is needed. If vmsep feature is
enabled, include only qemu-system-i386 package in the image.

Signed-off-by: Sai Hari Chandana Kalluri <chandana.kalluri@xilinx.com>