petalinux-image-full.inc: Refactor to avoid vitis-ai and ultra96 packages
Issue: CR-1061767
Issue: CR-1061749

Refactor the way the IMAGE_FEATURES are enabled to ensure that they are
only turned on for the correct soc family.

The only difference in output from before and after this change is to
the versal configuration.  The following items are no longer defined
when building an image for versal:

 petalinux-96boards-sensors
 petalinux-ultra96-webapp
 petalinux-vitisai

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>