gn-native: requires C++-14 and "string_view"
Issue: CR-1061749

gn-native is used to configure chromium.  It is based on existing
chromium sources, and thus requires C++-14 and specifically the
"string_view" C++ header.

Add a specific check for this header to ensure that we don't fail
during compilation, but instead fail during parse.  This provides
the user with a much better expereince in the case of failure.

Included with this bbappend are instructions for CentOS / RHEL 7
to create a working host configuration.

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>