image-xlnx-xen:config relative path for -d option
This patch provides the config file realtive path for -d option.  As this is
required because when user specifies the custom xen.cfg then the full path is
appended in the generated boot files which cannot be shared across machines.

so image-builder has -d option to take config relative path and generate the
outputs for sharing.

Signed-off-by: Swagath Gadde <swagath.gadde@xilinx.com>
Signed-off-by: Sai Hari Chandana Kalluri <chandana.kalluri@xilinx.com>