image-xlnx-xen: Rework to keep everything in the workdir
Issue: CR-1059795

Remove the python fragments from linux_xlnx.  Instead we move to a bbappend
approach.

No longer need to delete the do_deploy_setcscene task, as all actions happen
inside of the workdir, and then deploy to the DEPLOYDIR.  With the dependencies
if anything changes the do_deploy will be re-run, otherwise sstate-cache will
be used to depoy the contents into the deploy/images.

We need to explicitly list the do_deploy dependencies as well, because we
rely on a few other items to be present before we can deploy properly.

Also deploy the matching rootfs.cpio.gz to ensure the one we used is
deployed under the name we're referencing it as.

Finally, we add the new files to the PACKAGES_LIST so that they will be
deployed if the plnx-deploy class is enabled.

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>