xen: Ensure that do_deploy will have necessary dependencies
Issue: CR-1059795

If the sstate-cache is used, it's possible to get to do_deploy without
the uboot-mkimage being available.  We need to explicitly call out this
dependency if we intend to use it.

Additionally if sstate for all prior tasks are available the install
directory, ${D}, might not be available so we need to explicitly list the
do_install task as a dependency as well.

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>