Various QT Packages end up with direct links to libGL providers
Issue: CR-1082769

libGL is provided by either mesa or libmali.  So we need to ensure that
items that have these linkages declare specific dependencies and since
libmali is a SOC_VARIANT_ARCH packge, they must also be declared as such.

If we don't do this, the package will be generated as aarch64, but won't
be installable as the dependency on libmali can't be satisfied unless
the user has libmali, which is only available on zynqmpev and zyncmpeg.

Cherry-pick: dc0392c from 2020
Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>