layer.conf: Adding default package uri and archs
Adding default package uri and archs so if package-management is
enabled, these defaults will be used when creating the .repo file.

PACKAGE_FEED_ARCHS is set to the superset of all archs available in the
xilinx feed to be able to use one .repo file no matter which board, and
the .arch file will be modified to correctly pick packages from the
generic feed.

Signed-off-by: Jaewon Lee <jaewon.lee@xilinx.com>