Allow for a more secure initial device configuration
In conjunction with a configuration that does NOT have debug-tweaks enabled:

EXTRA_IMAGE_FEATURES_remove = 'debug-tweaks'

This will allow a user to be added to the system (with or without initial
password), force a password change on first login, and allow us to add that
user to a special /etc/sudoers.d/99-petalinux file.

If the new user does NOT have a password set (i.e. it's set to blank) then
removal of debug-tweaks will prevent remote login to that user, the user will
be required to use the console to set the initial password.

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>