Add support for PR Service within an eSDK
Issue: SSW-773

The ability to package PR Service data in the eSDK was missing upstream.
This required it to be implemented within both the class, populate_sdk_ext
and the eSDK install program files/ext-sdk-prepare.py.

Override the default oe-core versions by using a prepend within the
layer.conf file.  Also capture the path to the layer, so we can find
the replacement files/ext-sdk-prepre.py as well.

Finally, we unconditionally enable PR service when building petalinux.
The user can override this as necessary, but the default ensures a
local PR service database.

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>