xen_git.bbappend: Resync to latest meta-virtualization gatesgarth
The current version of gatesgarth changes the way the do_deploy should be
implemented.  It requires the install steps to be done from ${B} (build)
directory, and NOT ${D}.

It also uses slightly different naming then we expect, so add a
compatibility link for the time being.

The setup of the FLASK_POLICY_FILE has been moved to xen-tools recipe.

We also remove the _setscene deletions, as the Yocto Project build
requires this in order to generte an eSDK.  You are no longer allowed
to skip ANY setscene tasks in a configuration involving the eSDK.

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>