conf/layer.conf: Raise priority of the layer to 6
Issue: CR-1097824

The petalinux layer has the same priority as the oe-core layer, this can
cause an odd behavior where different packages of the same name may be
picked up from different layers based on minor configuration changes.

Since we need to use the local version of kernel-devsrc.bb, we need to
ensure this layer can replace the version in oe-core.

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>