linux-xlnx_2021.1.bb: Setting DT_SEARCH_ARG to look for dtbs only
With fpgamanager distro feature enabled, pl-final.dtbo, pl.dtbo, and
system-top.dtb are created. with recent changes in core, kernel-fitimage
class now considers dtbos when creating the fit image. additionally the
artifacts in EXTERNAL_KERNEL_DEVICETREE are sorted for reproducibility,
so pl-final.dtbo is picked to be the default dtb which causes issues on
bootup of the image.
Note: there is a dtbcount mechanism to be able to manually set a
numbered default, but this is not reliable (as well as raises error:
"Properties must precede subnodes" in our case)

Setting DT_SEARCH_ARG to go back to only processing dtbs when putting
together the fitimage.

Signed-off-by: Jaewon Lee <jaewon.lee@xilinx.com>