image-update: Fix compilation
image-update needs to be compiled and linked with the proper system cflags
and ldflags.  Adjust the calling of the makefile to ensure they are used.

Additionally set the package arch to be SOC_FAMILY specific so the
resulting package(s) can not be used on other systems.  This preserves
the COMPATIBLE_MACHINE exclusion.

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>