wpa-supplicant: don't launch dbus service and add -u option for wlan0
Two instances of wpa_supplicant are running simultaneously, one is started by
the if-pre-up.d script, the other by the dbus service. This results in a
conflict where the wlan0 adapter deauthenticates with error code
"3=DEAUTH_LEAVING". To work around this issue, comment the Exec line in the dbus
service file so wpa_supplicant does not get executed and add the -u switch which
enabled the dbus control interface as an additional argument to the if-pre-up.d
script. As a result, only one instance is running and wlan0 authenticates
properly.

This commit is cherry-picked from
https://github.com/Avnet/meta-avnet/commit/ab2ec6d87f8745f2b12b922564f766bbe1f5044d

Signed-off-by: Christian Kohn <christian.kohn@xilinx.com>
Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>