wpa-supplicant: Remove custom wpa-supplicant script
Unfortunately, the modified wpa-supplicant script does not behave as intended
when started as a daemon during init. Adding the -u flag causes the process to
terminate when run during init which in turn causes the link up to fail. Keep
using the default script without -u as a workaround which means wpa-supplicant
is started without dbus support.

This fixes b64547c.

Signed-off-by: Christian Kohn <christian.kohn@xilinx.com>
Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>