petalinux-features: Enable both sysvinit and systemd support
This enables packages to support both sysvinit and systemd.  The image
configurate is what determines which initscript system is enabled.

BACKFILL needs to be used, otherwise it can trigger obscure issues when
just setting sysvinit or systemd alone.

petalinux-image-minimal, for zynqmp-generic:

  systemd - 170136k
  both    - 170264k

comparison between switching to sysvinit and both enabled,
petalinux-image-minimal, for zynqmp-generic:

  sysvinit - 130244k
  both     - 135200k

The difference is down to udev.  In the normal sysvinit configuration
"eudev" is used, while in the both case a systemd based udev is used
instead.  This is newer and more feature rich, but has slightly more
dependencies.

Having both also enables a single binary distribution package set
that will work with either sysvinit or systemd, as well as uses
the same udev so we only have a single set of potential bugs to
resolve.

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>