petalinux-image-common: Set default systemd target
If either the x11 or weston petalinux packagegroups are included in the build
via IMAGE_INSTALL, then set the systemd default target to graphical.target else
to multi-user.target. Otherwise, the display-manager service is not started
automatically even if it is enabled as it is not requested by any other service.
This setting only impacts images with systemd enabled.

Signed-off-by: Christian Kohn <christian.kohn@xilinx.com>
Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>