linux-xlnx_%.bbappend: Adding required kernel configs for systemV
New udev with systemv is throwing below error but boot went on.
"/proc/sys/kernel/hotplug: No such file or directory"

If we enable CONFIG_UEVENT_HELPER in kernel then the issue is not seen
and community is suggesting to go with systemd and they obsoleted this
feature in udev.

So for zynq as we are with systemV enabling the above config in kernel
default config.For aarch64 as we are having default systemd enabling this
kernel config only in systemv enabled case.

Signed-off-by: Varalaxmi Bingi <varalaxmi.bingi@xilinx.com>
Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>