start-jupyter*: Don't wait for IP in start-jupyter script
systemd allows a service to be started after the network is up, hence there is
no need to actively wait for an IP address to get assigned. Create a function
named wait_for_ip in the sysV init script to retain this functionality for sysV
but remove the wait loop from the wrapper script which is called from either
sysV or systemd.

Signed-off-by: Christian Kohn <christian.kohn@xilinx.com>
Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>
(cherry picked from commit 49960b270177948d1404cc3f889668d35303858b)
Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>