Add swtpm and libtpm to the list of packages
swtpm and libtpm is to be included in the list of packages to be installed in
the generated rootfs. The reason being the user is expected to run swtpm from
the console which will interact with another application (eg qemu) to execute
the tpm commands received and send the response.

Signed-off-by: Ayan Kumar Halder <ayankuma@amd.com>
Signed-off-by: Sandeep Gundlupet Raju <sandeep.gundlupet-raju@xilinx.com>