petalinux-initramfs-image: Remove ${MACHINE_ESSENTIAL_EXTRA_RDEPENDS}
MACHINE_ESSENTIAL_EXTRA_RDEPENDS is used to specify everything required to
boot the device.  The petalinux-initramfs-image is often used as an early
ramdisk and thus may be included (indirectly) in the RDEPENDS, creating
a depenency loop.  Also the files provided by the packages in the RDEPNDS
are not necessary in a ramdisk image, and just waste space.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>