petalinux-version: Fix eSDK XILINX_VER_BUILD
When an official build happens, XILINX_VER_BUILD is replaced with a hardcoded
version.  However, a build from git uses the git commit, via
METADATA_REVISION.  METADATA_REVISION is only set when building from a git
repository, otherwise it's set to '<unknown>'.  This causes problems with
some SDK creation, due to the < and > characters.  Search for this and
replace it with 'unknown' will avoid the problems with shell redirection.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>