meta-xilinx-toolchain: Add runtime-relocatable (-r) option
Add special processing that will create a shell wrapper that calls the
dynamic loader, with the correct library path for each binary.  The
binaries are then renamed with '.real'.  This will work in most cases,
but some programs that look up the executable names and try to call
them directly will likely fail.

The meta-xilinx-toolchain SDK, when installed will work the same way
as a Yocto Project SDK by default.  When adding the '-r' (runtime
relocation) optional will invoke the additional scripting.  While the
compiler can be invoked directly, it is highly recommended the user
continue to source the environment script to get the correct mandatory
CC, LD and AS flags.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>