searche2fs:update search delay to 2 seconds
we have 1.5 second delay in searche2fs script while switching to full rootfs but since SD is on USB.
so within 1.5 second  SD device is not up and unable to mount rootfs, and its landing in tiny rootfs.
so removed loop and increased the delay to 2 seconds to mount rootfs from SD on USB device.

Signed-off-by: Swagath Gadde <swagath.gadde@amd.com>
Signed-off-by: Mark Hatle <mark.hatle@amd.com>