plnx-baremetal-toolchain-scripts: Adjust the environment files for multilibs
The special multilib PetaLinux/Vitis toolchain only has a single set of
binaries for the toolchain components.  Use a customer set of scripts to
define the environment files to call the primary named multilib binaries,
but with the multilib specific arguments.

This also introduces the meta-environment.bbappend which activates this new
code.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>