petalinux-initramfs-image.bb: Update IMAGE_FSTYPES
Create python anon function to update the IMAGE_FSTYPES with
INITRAMFS_FSTYPES as we are using :<override> for IMAGE_FSTYPES in .conf
file which will override the value defined in petalinux-initramfs-image.

Signed-off-by: Raju Kumar Pothuraju <rajukumar.pothuraju@amd.com>

Added the IMAGE_FSTYPES:forcevariable to ensure that variable hashes will
be maintained.  This may not be required, but will help prevent the case
where changes to INITRAMFS_FSTYPES won't cause a rebuild.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>