core-image-*-initramfs.bbappend: Add a python call to avoid override issues
Create python anon function to update the IMAGE_FSTYPES with
INITRAMFS_FSTYPES as we are using :<override> for IMAGE_FSTYPES in .conf.

Using the forcevariable as a backup to ensure hash calculation to ensure
changes to INITRAMFS_FSTYPES trigger a rebuild.

Note, these likely won't be used be end users, but it is useful for
testing purposes.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>