Replace initramfs bbappend with plnx initramfs
The petalinux-initramfs-framework bbappend replaces a number of the
upstream modules with petalinux specific modified versions making it
difficult to use the upstream framework unmodified.

This change replaces the bbappend with a petalinux specific version of
the framework which inherits the unmodified modules from the upstream
recipe but provides modified or additional petalinux specific modules.

petalinux-initramfs-image is also updated to use the new
plns-initramfs-framework.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>