petalinux-features.conf: Use two steps on indirection for INIT_MANAGER
The INIT_MANAGER must be set with '='.  This is how poky does it, and we need
to do similar.  If the user wants, they can do INIT_MANAGER:forcevariable = ...
but this is not recommended.

Instead we use the new various (previously defined) INIT_MANAGER_DEFAULT
which the user can override in their local.conf or similar context.

Previously, we used overrides for microblaze and zynq.  This prevented the
user from overriding INIT_MANAGER_DEFAULT, so we add yet another level of
indirection to define the 'petalinux' default, and then default the new
INIT_MANAGER_DEFAULT to the petalinux specific one.  This preserves the
ability for the user to override the value with one of their choosing.

Note, only sysvinit and systemd are tested, other values may work but have
not been tested.

Signed-off-by: Mark Hatle <mark.hatle@amd.com>